split k8s execuor file into logically independent files
